Refactor carousel state
