+++ react-carousel/src/plugins/rtl.js
@@ -1,15 +1,15 @@
 import { useEffect } from 'react';
 import { useRecoilState, useRecoilValue } from 'recoil';
 
-import {
-  carouselValueState,
-  slideMovementState,
-  slidesState,
-  trackStylesState,
-} from '../state/carousel';
 import { pluginNames } from '../constants/plugins';
 import STRATEGIES from '../constants/strategies';
 import clamp from '../tools/clamp';
+import { slidesState } from '../state/atoms/slideAtoms';
+import {
+  carouselValueState,
+  slideMovementState,
+  trackStylesState,
+} from '../state/atoms/carouselAtoms';
 
 import './rtl.scss';
