+++ react-carousel/src/state/atoms/carouselAtoms.js
@@ -1,0 +1,43 @@
+import { atom } from 'recoil';
+
+export const slideMovementState = atom({
+  key: '@brainhubeu/react-carousel/slideMovementState',
+  default: {
+    clicked: null,
+    dragStart: null,
+    dragOffset: 0,
+  },
+});
+
+export const transitionEnabledState = atom({
+  key: '@brainhubeu/react-carousel/transitionEnabledState',
+  default: false,
+});
+
+export const carouselWidthState = atom({
+  key: '@brainhubeu/react-carousel/carouselWidthState',
+  default: 0,
+});
+
+export const trackWidthState = atom({
+  key: '@brainhubeu/react-carousel/trackWidthState',
+  default: 0,
+});
+
+export const trackStylesState = atom({
+  key: '@brainhubeu/react-carousel/trackStylesState',
+  default: {
+    marginLeft: 0,
+    transform: 0,
+  },
+});
+
+export const carouselStrategiesState = atom({
+  key: '@brainhubeu/react-carousel/carouselStrategiesState',
+  default: [],
+});
+
+export const carouselValueState = atom({
+  key: '@brainhubeu/react-carousel/carouselValueState',
+  default: 0,
+});
